Replace fluttertoast with toastification
Migrates the application's toast notifications from the `fluttertoast` package to `toastification`.

This change involves updating the `showCustomToast` function in `lib/core/utils/funcation.dart` and the `AppToast` class in `lib/core/widget/app_toast.dart` to utilize the new `toastification` API.

The `pubspec.yaml` and `pubspec.lock` files have been updated to reflect the removal of `fluttertoast` and the addition of `toastification`.